Improve the loading of kdenlive 0.5 files (file format version 0.6?), by constructing a non-empty kdenlivedoc instance in the files, that gets its information from the new westley entry. This means that the clips shows up on the timeline and in the project tree. However, in order to avoid a crash in the audio thumblines, two checks for channels >= 1 has been added. Transitions and other stuff appears to show up, but the user should probably delete most transitions, as I've had a number of crashes when trying to edit them. However, I can now load a 4 minute 103 clip kdenlive 0.5 project, and save it in the new format.
I am committing this after a check with jmpoure on the mailing list,
but if the audio channel stuff is wrongly implemented, I will have to revert
it, of course, and instead either get the channels from the old
kdenlivedoc instance, or simply require of the user that he disables
the audio thumbnailing while loading a kdenlive 0.5 document.

Also, there is some ifdef'ed code to order the tracks and keep the
folder structure. I am still working on those, but because I got a bit
lazy, it is easier for me to commit them with this patch, than having
to reedit again. A later patch will clean this up.

svn path=/branches/KDE4/; revision=2615