Fix hang on exit somewhere strange inside Qt on OS X.
This fixes it on OS X, and stil works for me on Linux. Note that
explicitly deleting the window object on Linux causes a crash. By
setting a breakpoint in ~MainWindow, I see that before this change, on
OS X, the dtor is not called. Meanwhile, something in the GUI is likely
still using Qt during/after app.exit() (it was hanging in QtColor, but
the backtrace did not lead back to Kdenlive source). Howevever, on
Linux, without explicitly deleting the window object, the debugger still
stops on the breakpoint inside the dtor! Might this be dependent on Qt
version?

svn path=/trunk/kdenlive/; revision=4833