Audio alignment works.
This is at the moment still a brute-force attack (correlation
currently requires O(n²) time) but /works/ and is reasonnably
fast for clips with reasonnable length (not too many minutes).

Later the correlation can be replaced with an FFT version which
would only require O(n log n) time.

Example usage of the executable:
$ ./audioOffset mainVideo.avi someSound.wav

This will e.g. tell you that someSound.wav needs to be shifted
by 42 frames in order to be aligned. The MLT profile can be
specified with --profile=<mlt profile>.