Load the (effect) parameter types as plugins.
This required the introduction of a library to be able to access
effectsystem base classes and used widgets.
The library "kdenlivecore" is currently in src/core/ but might at a
later point of refactoring be moved up to src/ when it contains major
parts of the source to enable a powerful plugin system.