First work on new project tree ("Bin").
A data -> model -> view separation is used to make it easy to add new
views in the future. The data which consists of clips and folders has a
common base class AbstractProjectItem to make data -> model interaction
and loading of documents straight forward. Inside the document file the
items are now stored in a more hierarchical way to represent this object
structure.