Fix jittering issues when sending GL textures cross-thread.
When rendering into a texture on one thread and rendering _from_ it on another,
we need to properly synchronize the client state. Hook onto the OpenGL fences
that MLT set to properly wait for the rendering commands to be sent to the GPU,
which fixes the jittering we'd otherwise see, due to rendering old frames still
left in the textures because the new ones are not ready yet.