Configuration dialog UI improvement: show action texts assigned to jog buttons without "&"; easier to read; sorts according to user expectations
This proposed patch improves the configuration dialog UI for the Jog/Shuttle tab: the texts for actions that can be assigned to buttons of a jog/shuttle device previously contained "&". These ampersands cause action texts to be sorted in an unintuitive way that makes finding the correction action hard. This patch removes all ampersands from the action descriptive texts, so that the resulting action list is sorted according to user expectations. In particular, leading ampersands don't cause action names to get sorted into two sets: one going from &A-&Z and another one going from A-Z. With this patch, there is only a single sorted list A-Z without any distracting ampersands.

On a sidenote: Kdenlive up to 0.9.10 already removed the ampersands, only later they got in again during transition to KF5.

REVIEW: 125548