Timeline incorrectly checks track index bounds crash
Class Timeline incorrectly checks the bounds for track indices:
the upper bound is checked against greater than track.count(),
where in fact the indices run from 0 to count-1.

However, this masks the underlying problem with converted Kdenlive projects:
when adding new tracks, the index gets out of bounds,
so this needs to be fixed too.
This patch just fixes the crash in itself but not the underlying problem
of handing in an out-of-bounds index to these functions to be fixed.

REVIEW: 125707