fix clipping of audio signal view db marks
In vertical mode the audio signal view clips some of its audio level marks.
The reason is that the drawing method does not take
into consideration the actual font/text dimensions.
This proposed patch fixes the incorrect clipping for vertical mode, that is,
when the audio signal view is taller than wide.
It calculates the actual width of the longest db level mark,
which it assumes to be the last mark in the db level list,
instead of using the present hardcoded "fudge" values.

REVIEW: 125708