Clearer icon design for effect split monitor view button
both light/dark breeze compatible icon
Currently, the effect split monitor view button shows a "shuttle door".
Proposed here is an improved design that instead communicated to users the effect meme in form of a star.
To signal split view mode, the only the left part of the star is visible.

The proposed icon design consists of both light and dark breeze icons.
In addition, an optional icon is provided that could be used when in effect split mode to signal how to switch back to effect view mode.