Allow building on ubuntu armhf (OpenGL!=OpenGLES)
BUG: 350061

applying:
kdenlive-15.08.3-define-gl-unpack-row-length-if-missing.patch
kdenlive-15.08.3-only-delete-opengl-3.2-core-functions-on-windows.patch

The two patches fix all serious warnings in the glwidget.cpp file
and make kdenlive-15.08.3 compile.

GL_UNPACK_ROW_LENGTH is part of OpenGL ES 3.2, but not of OpenGL ES 2.0.
This it is defined in different headers (gl3.h) in OpenGL ES 3.2 and as an
extension with _EXT suffix (gl2ext.h) to OpenGL ES 2.0.
I chose the route to fallback to the _EXT definition, in case the regular one
is unavailable.

The other patch fixes a warning about calling operator delete on
QOpenGLFunctions_3_2_Core (its class definition is incomplete).
It is only ever instantiated on Windows anyway,
so I took the liberty to also guard the deletion with Q_OS_WIN.