Remove Debian menu file and its xpm icon
The menu file was added back in 2009 at the request of the Debian
packager; recently [1], the Debian technical committee ruled that the
Freedesktop menu is now the preferred one, and a menu file must not be
installed if a .desktop file is provided (and kdenlive provides one).
The menu file has not been installed anymore since few months [2] as a
result of that.

Hence, drop the menu file and the xpm icon for it, as no more useful.

[1] https://lists.debian.org/debian-devel-announce/2015/09/msg00000.html
[2] https://packages.qa.debian.org/k/kdenlive/news/20151026T180601Z.html