src/dvdwizard/dvdwizardmenu.cpp: do not show "grid" in output
The dvd wizard supports a "grid" to help positioning elements
and it displays small dots so that users can easily position and
align elements in the dvdwizard. This is nice but those are
currently also visible in the output in the dvd menu.

This is fixed by disabling the grid for the rendering.
BUG: 377256