Fix opening relative paths from the command line
The code didn't work as intended, and somewhere something was calling
QUrl::fromUserInput() when adding to recently used documents. And that
helpfully adds "http:" as scheme if none is set (e. g. if opening a
relative file). So a lot of things broke.

Test Plan: Opening relative paths from the command line works again

Differential Revision: https://phabricator.kde.org/D28628