Update org.kde.kdenlive.desktop to inherit correct window title and icon, instead of lowercase title and generic icon
Tested with Kdenlive 20.04.0 on latest Arch. The application title and
icon were correct in the GNOME Applications view, but upon opening
Kdenlive, the window title was "kdenlive" and the window icon did not
match my system Kdenlive icon, which is provided by Papirus (but this
applies to other themes that provide a Kdenlive icon). Adding a startup
WM class ensures the window title is capitalised the the window icon
matches the user's application icon.