PACKAGING CHANGE!! Introduce PACKAGE_TYPE envvar
When Kdenlive runs inside a sandbox enviroment like with Appimage,
Flatpak or Snap it needs to resolve some paths rekursive. This is
currently detected by whether a custom config file was specified or not.
This patch changes the behaviour. The package type is now detected
through the envvar PACKAGE_TYPE. You can set it to "appimage", "flatpak"
or "snap" to let Kdenlive run in sandbox mode.