Lower proxy rendering priority
This adds a Kdenlive setting to lower the priority of the proxy rendering QProcess. This helps keep the main UI responsive when proxies are rendering.

One problem is that setting the niceness this way only works on Unix. I'm not sure how to do the same on Windows so any tips there would be appreciated. (I also can't build Kdenlive on Windows.)