Fix Insert Zone to Bin out point off by 1
If the user sets in=100, out=101 and select Insert Zone in Project Bin,
the clip is created 1 frame longer than it should be. This is especially
noticeable when creating a subclip when the out point is the last frame
in the clip, resulting in a broken subclip that cannot be added to the
timeline.

BUG: 455883