Make monitor zone out be last frame number
This should fix a mismatch in behavior between the monitor and the rest
of Kdenlive. The out frame number should be the last shown frame of a
clip. E.g. in=100, out=200 has 101 frames. Previously, the monitor
treated the out point as the frame number *after* the last shown frame,
causing an off-by-1 bug when taking in/out points from the monitor and
using them in the rest of the program.

Changing the definitions to be consistent across Kdenlive will reduce
the number of places in the code that need a +1 or -1 to adjust the
frame numbers. Note that duration calculations will still need to offset
by 1, e.g. duration = out - in + 1.