Guess subtitle encoding before importing
Since many subtitle files are not UTF-8, we need to guess the encoding
of the file before reading it. For example, SubRip's default encoding is
Windows-1252 (according to Wikipedia).

This also adds KF5 Codecs as a dependency in order to use KEncodingProber.

Future work could be done to allow the user to select the encoding in the import dialog. Currently there is no way to manually select the encoding if it's not guessed correctly, but this should at least be an improvement over only supporting UTF-8.

BUG: 456871