Fix filtering TreeItem lists by non-ASCII strings
AssetFilter tries to normalize the TreeItem's text by removing any
punctuation and symbols, but it was done using a simple [^a-zA-Z0-9\s]
character class which removed any non-ASCII characters, breaking the
search for many non-English languages. Replacing this with [^\w\s]
didn't work, but iterating through the string and using
isLetterOrNumber() does work.

BUG: 432699