[Kdenlive Settings] Remove "preview volume" option
This was used in a different way back in time, but now it does not make
sense to have this option in the settings anymore. The UI element in the
settings is a duplicate of the volume slider in the monitors hamburger
menu, but in the current implementation both options are not properly
syncronized.