Warn user when low on memory
In the Render Job Queue dialog, show a warning to the user when the system is low (< 128MB) on memory. If the system hangs due to running out of memory, this will help the user understand why. The message shown is, "Less than [xx]MB of available memory remaining." Uses KMemoryInfo if available (KF >= 5.95), otherwise Windows, Linux, Mac, and FreeBSD supported (FreeBSD code copied from KMemoryInfo). Should be a no-op on other OSes.

Fixes #1599