Ensure there are always guide categories available
If there is not at least one category this will cause problems in
several parts of the app. We do not allow to delete all categories, but
if for whatever reason the project file does not contain any, we will fix
that on project open and add a default set of categories.