Audio Capture Subdir
This introduces a new option to the environment config to allow captures
to the stored in a subdirectory of the project folder on disk, rather
than only in the root.

To go along with that this also cleans up the current confusion in the
recording classes around which folder to put stuff in, putting all
captures into the capture folder, rather than splitting them between
audio captures and video captures, where video captures end up in with
the renders. Unless a custom folder is set for audio, then they end up
in there *with* the audio captures.

This is cleaner. Captures all go in the capture folder.
And we make the folder if it's pointing to a non-existent folder. This
was previously theoretically possible if someone set a custom folder
that didn't exist, but that would be a weird thing to do.

Now, though, it's common for people setting subdirs, because each new
project will probably not have a capture dir before their first capture.

Also, now that there's two options that require follow-up info, I made
it so when you don't have "custom folder" chosen, rather than showing
the custom path but having it disabled, I just hide it. And then added
another field that shows up to ask the subdir when you've chosen that.

In doing this, I've taken the `getProjectFolderName` method, which was
previously effectively two independent methods toggled by a boolean, and
made them two fully different methods.

Closes #1395