Don't Highlight Newly Recorded Audio in the Bin
When an audio capture is recorded on an audio record track, it moves the
"cursor" to the bin, and highlights the clip it just recorded.

I find this disruptive, because often I'll have a clip in the bin that
I'm pulling zones out of, then putting them into the timeline, then
recording some voice-over, then pulling another zone, etc.  Or I'll be
pulling clips that I've previously organized into folders of similar
clips.

In either case, suddenly having my bin focus get moved to some other
area, and having it focused on this thing I just recorded, is never what
I want.  If I want to listen to it, I should listen to it in context, on
the timeline where it was just inserted. I don't need to look at the
clip. And if I really want to, I can always right click on the new
thing in the timeline and use that to find it in the bin like any clip.

By keeping my "focus" on the timeline, it also means I can continue to
use keyboard shortcuts to move around my newly inserted recording, like
jumping to the end of it or the beginning, or the next clip, etc.

I have written this PR as though everyone agrees with me, and that no one wants the current behaviour, because it is simpler. But if it turns out that's not true, and there is a workflow other people use that requires this, we could make a more complicated version of this PR that introduces a new setting somewhere and then does these things conditionally rather than just deleting them entirely.