Improve AV1 NVENC export profile
This improves the AV1 NVENC export profile introduced in https://invent.kde.org/multimedia/kdenlive/-/merge_requests/409

The current values produce large files and the quality slider is limited to only a small range (`15` - `45`) of all possible QP values. Therefore even when setting the quality slider to 0% the output file is still big and it is not possible to compress the video to a more reasonable size

The AV1 NVENC encoder supports a QP range of `-1` - `255`.
This MR fixes the above issue by exposing more of this range to be able to output smaller files

I also changed the default value to one which outputs a bit smaller files compared to H264 with its default CQP. I think this is what most users expect considering that AV1 is in most cases a more efficient codec

There are more possible improvements to investigate (CQ mode instead of CQP, Spatial AQ & Temporal AQ, testing on more content to find better values overall and better naming of the profiles...)
but for now this should be fine and gets the profile in a usable state for most situations