Reassemble proxy profile elements in the correct order after validation
The external proxy preset strings consist of sets of 6 elements:
path clip->proxy;proxy prefix;proxy suffix;path proxy->clip;clip prefix;clip suffix
appended sequentially for each 'profile' that the preset describes.

When we import those into the proxy profiles edit dialog, we deconstruct
that and regroup them so that all the clip->proxy paths (etc.) are together
in a single QLineEdit widget for each element type.

Which means we need to reverse that process and re-interleave them in
profile groups after validating those fields.  If we just concatenate
the contents of the edit widgets all the elements end up in the wrong
order.

BUG: 485356