docs: update README
I'd like to propose some changes to the README:

After brief welcome and introduction what this project is I would like to point regular users to the website. It contains the best and most up-to-date info for everything except dev/coding documentation.

Everything after that would then focus only on developers / potential code contributors.

I imagine the README to be the landing page we forward people to from the website that are interested in contributing code / hacking on the project. (On the new website this would be from the contribute page https://invent.kde.org/websites/kdenlive-org/-/merge_requests/22)