Fix Vectorscope zooming artifacts on high zoom level
When zoomed (gain > 1.0) we did not plot additional pixels in the zoomed region so those remain as fill color producing visible artifacts making it hard to see anything on the scope.

Fixed by interpolating these additional in-between pixels so we get a smooth continuous view without artifacts.

Using QImage SmoothTransformation method here for a smooth/blurry look. With FastTransformation we'd be 2x faster but compared to the time it takes to draw the vectorscope overall the scaling time doesn't really matter.

Testing on a 1080p clip:

Before:

**101_000 us**

![no_transformation](/uploads/18daf0a6fc13c5c661ae46f12dc29346/no_transformation.png)

After:

**102_000 us**

This includes scaling with SmoothTransformation which itself took **700 us**.

![smooth_transformation](/uploads/83bccc2d15e4aa004301534a030dba5c/smooth_transformation.png)

Interpolation Method:

I went with SmoothTransformation but other options would be possible.

FastTransformation would be faster at around **300 us**. But given the time it takes to draw the unscaled Vectorscope this gain is pretty much negligible.

This is how it would look using FastTransformation instead of Smooth:

![fast_transformation](/uploads/1708b1e95628efa36d53543635c580a5/fast_transformation.png)

This is what it looks like in Davinci Resolve: https://youtu.be/m1F9TJzfo1s?feature=shared&t=483

Too me this looks blurry so I guess they use Bilinear/Smooth interpolation.

@emohr was in favor (via Chat) of following Davinci here.