Fix ugly icons in Clip Properties panel
On my system (Arch Linux, Qt 6.9.1 KDE Frameworks 6.17.0) the Clip Properties Panel looks like this:

![ugly](/uploads/c3214eb38b1fb604ec5631774c231bbb/ugly.png)

I assume this not always looked like this so maybe this is a Breeze or Qt bug?

I found a workaround by overwriting the style sheet for QTabBar, this is how it looks then:

![pretty_overwritestylesheet](/uploads/0f353f92a6008250fdfda7f613e32f86/pretty_overwritestylesheet.png)

Another Alternative would be to set DocumentMode to false then these ugly gray overlays will not be shown either. Then it would look like this on my system:

![pretty_nondocumentmode](/uploads/91f69f9f924e879802fe63b7efa8780d/pretty_nondocumentmode.png)

I tested 24.12 AppImage and get the same result so it's probably not a recent change/bug in Breeze/Qt)
The docs about QTabBar DocumentMode mention MacOS so maybe this is specific to Linux? :thinking:  Can you check if you can reproduce this issue on Windows/MacOS?