Drop unused v4l dependency
v4l dependency was used when we implemented a webcam capture feature. This feature has been dropped for some time and is not working anymore.
So there is no point in keeping the dependency. If we one day decide to reimplement it, we should probably try to do it through QtMultimedia, not through v4l.
CCBUG: 467391