Switch Kdenlive layouts to separate json files for each layout.
The Kdenlive layouts are currently stored in a config file (`kdenlive-kddlayoutsrc` in ~/.config).
This is rather ugly as it means we have full layout data in the form of:

`kdenlive_editing={"allDockWidgets": [{"lastCloseReason": 0.. }`

With lines having more than 15'000 characters length.
This MR changes this to have each layout stored in its own json file in `~/.local/share/kdenlive/layouts`.

The code change is rather long and maybe not perfect, but I think it would be great to make the switch now as we are introducing a new layout format with KDDockWidgets.