Remove option to fetch OTIO
Use a system installed version of OpenTimelineIO instead.

Now that we did a few releases with OpenTimelineIO already, distros are starting to ship it. With the recent release of OTIO 0.18.1 all major issues we had before are fixed.

For cases like development or if OTIO is not provided by the distro, it is not too heavy to build OTIO yourself.