[CI] Add third-party/kddockwidgets dependency
kde-builder uses .kde-ci.yml files to create a list of dependencies
for every package. With KDDockWidgets now buildable by kde-builder,
we can now add this dependency here to make it so the user doesn't
have to build KDDockWidgets themselves.

This change requires the following MR to be merged:
sysadmin/repo-metadata!660